# docs/swagger.yaml
basePath: /multicast/v1
definitions:
  basic.Message:
    properties:
      MessageHeader:
        additionalProperties:
          type: string
        type: object
        example : { "type": "TOCMULTICAST", "GroupId": "0"}
      Payload:
        items:
          type: integer
        type: array
    type: object
  multicastapp.Member:
    properties:
      address:
        type: string
        example : 127.0.0.0:90
      member_id:
        type: string
        enum: [0,1,2,3]
      ready:
        type: boolean
        example : true
    type: object
  multicastapp.Message:
    properties:
      MessageHeader:
        additionalProperties:
          type: string
        type: object
        example : {  "type": "TOCMULTICAST", "GroupId": "0"}
      Payload:
        items:
          type: integer
        type: array
    type: object
  multicastapp.MulticastInfo:
    properties:
      members:
        additionalProperties:
          $ref: '#/definitions/multicastapp.Member'
        type: object
        example : { "members": {
          "member1": {"address": "127.0.0.01:90", "member_id": 0, "ready": true  },
          "member2": {"address": "127.0.0.02:90", "member_id": 1, "ready": true  },
          "member3": {"address": "127.0.0.03:90", "member_id": 2, "ready": true  },
          "member4": {"address": "127.0.0.04:90", "member_id": 3, "ready": true  }}}
      multicast_id:
        type: string
        example : MYGROUP
      multicast_type:
        type: string
        example : BMULTICAST
      received_messages:
        type: integer
        example: 5
      status:
        type: string
        example: OPEN
    type: object
  multicastapp.MulticastReq:
    properties:
      multicast_id:
        type: string
        example : BASICGROUP
      multicast_type:
        type: string
        enum: [BMULTICAST,TOCMULTICAST,TODMULTICAST,COMULTICAST]
    type: object
  multicastapp.Response:
    properties:
      data:
        type: object
    type: object
  utils.DelivererNode:
    properties:
      nodeId:
        type: string
        example : 127.0.0.1:90
    type: object
  utils.Delivery:
    properties:
      deliverer:
        $ref: '#/definitions/utils.DelivererNode'
      m:
        $ref: '#/definitions/basic.Message'
      status:
        type: boolean
    type: object
info:
  contact:
    email: salvatimartina97@gmail.com
  description: This is a multicast API
  title: MULTICAST API
  version: "1.0"
paths:
  /deliver/:
    get:
      description: Get Deliver-Message queue of Group by id
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/utils.Delivery'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/multicastapp.Response'
      summary: Get Deliver-Message queue
      tags:
        - deliver
  /deliver/{mId}:
    get:
      description: Get Deliver-Message queue of Group by id
      parameters:
        - description: Multicast group id group
          in: path
          name: mId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/utils.Delivery'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/multicastapp.Response'
      summary: Get Deliver-Message queue
      tags:
        - deliver
  /groups:
    get:
      consumes:
        - application/json
      description: Get Multicast Groups
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/multicastapp.MulticastInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/multicastapp.Response'
      summary: Get Multicast Groups
      tags:
        - groups
    post:
      consumes:
        - application/json
      description: Create Multicast Group
      parameters:
        - description: Specify the id and type of new multicast group
          in: body
          name: request
          required: true
          schema:
            $ref: '#/definitions/multicastapp.MulticastReq'
      produces:
        - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/multicastapp.MulticastInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/multicastapp.Response'
      summary: Create Multicast Group or join in an existing group
      tags:
        - groups
  /groups/{mId}:
    delete:
      consumes:
        - application/json
      description: Delete an existing group
      parameters:
        - description: Multicast group id group
          in: path
          name: mId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/multicastapp.MulticastInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/multicastapp.Response'
      summary: Delete an existing group
      tags:
        - groups
    get:
      consumes:
        - application/json
      description: Get Multicast GroupInfo by id
      parameters:
        - description: Multicast group id group
          in: path
          name: mId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/multicastapp.MulticastInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/multicastapp.Response'
      summary: Get Multicast GroupInfo by id
      tags:
        - groups
    put:
      consumes:
        - application/json
      description: Start multicast group by id
      parameters:
        - description: Multicast group id group
          in: path
          name: mId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/multicastapp.MulticastInfo'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/multicastapp.Response'
      summary: Start multicast group by id
      tags:
        - groups
  /messaging/{mId}:
    get:
      consumes:
        - application/json
      description: Get Messages of a Group
      parameters:
        - description: Multicast group id group
          in: path
          name: mId
          required: true
          type: string
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/multicastapp.Message'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/multicastapp.Response'
      summary: Get Messages of a Group
      tags:
        - messaging
    post:
      consumes:
        - application/json
      description: Multicast a message to a group mId
      parameters:
        - description: Multicast group id group
          in: path
          name: mId
          required: true
          type: string
        - description: Message to multicast
          in: body
          name: message
          required: true
          schema:
            $ref: '#/definitions/multicastapp.Message'
      produces:
        - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/multicastapp.Message'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/multicastapp.Response'
      summary: Multicast a message to a group mId
      tags:
        - messaging
swagger: "2.0"